combine with existing project
