+++ frontend/src/data/alan/apa_2007_2017_data.js
