+++ frontend/src/data/alan/ncha_academic_issue_factors_data.js
