+++ frontend/src/data/alan/ncha_attempted_suicide_data.js
