+++ frontend/src/data/alan/ncha_felt_mentally_exhausted_data.js
