+++ frontend/src/data/alan/ncha_felt_overwhelming_anger_data.js
