+++ frontend/src/data/alan/ncha_felt_overwhelming_anxiety.js
